+++ pierky/arouteserver/commands/clients_from_euroix.py
@@ -14,20 +14,13 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import argparse
-import json
-import logging
 import sys
-try:
-    # For Python 3.0 and later
-    from urllib.request import urlopen
-except ImportError:
-    # Fall back to Python 2's urllib2
-    from urllib2 import urlopen
 import yaml
 
 from .base import ARouteServerCommand
 from ..config.program import program_config
 from ..errors import EuroIXError, EuroIXSchemaError
+from ..euro_ix import EuroIXMemberList
 
 class ClientsFromEuroIXCommand(ARouteServerCommand):
 
@@ -36,8 +29,6 @@ class ClientsFromEuroIXCommand(ARouteServerCommand):
                     "of EURO-IX JSON file.")
     NEEDS_CONFIG = True
     
-    TESTED_EUROIX_SCHEMA_VERSIONS = ("0.4", "0.5", "0.6")
-
     @classmethod
     def add_arguments(cls, parser):
         super(ClientsFromEuroIXCommand, cls).add_arguments(parser)
@@ -60,8 +51,11 @@ class ClientsFromEuroIXCommand(ARouteServerCommand):
         parser.add_argument(
             "ixp_id",
             type=int,
+            nargs="?",
             help="The numeric identifier used by the IX to identify the "
-                 "infrastructure for which the list of clients is requested.")
+                 "infrastructure for which the list of clients is requested. "
+                 "If not given, a list of IX's infrastructures will be "
+                 "printed.")
 
         parser.add_argument(
             "--vlan-id",
@@ -88,232 +82,17 @@ class ClientsFromEuroIXCommand(ARouteServerCommand):
             default=sys.stdout,
             dest="output_file")
 
-    @staticmethod
-    def clients_from_euroix(data, ixp_id, vlan_id=None,
-                            routeserver_only=False):
-
-        def check_type(v, vname, expected_type):
-            if expected_type is str:
-                expected_type_set = (str, unicode)
-            else:
-                expected_type_set = expected_type
-
-            if not isinstance(v, expected_type_set):
-                if expected_type is int and \
-                    isinstance(v, (str, unicode)) and \
-                    v.isdigit():
-                    return int(v)
-
-                raise EuroIXSchemaError(
-                    "Invalid type for {} with value '{}': "
-                    "it is {}, should be {}".format(
-                        vname, v, str(type(v)), str(expected_type)
-                    )
-                )
-
-            return v
-
-        def get_item(key, src, expected_type=None, optional=False):
-            if key not in src:
-                if optional:
-                    return None
-                raise EuroIXSchemaError("Missing required item: {}".format(key))
-            val = src[key]
-            if expected_type:
-                val = check_type(val, key, expected_type)
-            return val
-
-        def new_client(asn, description):
-            client = {}
-            client["asn"] = asn
-            if description:
-                client["description"] = description.encode("ascii", "replace")
-            return client
-
-        def get_descr(member, connection=None):
-
-            res = []
-
-            for info, k in [("AS{}", "asnum"),
-                            ("name '{}'", "name")]:
-                try:
-                    res.append(info.format(member[k]))
-                except:
-                    pass
-
-            if connection:
-                for info, k in [("ixp_id {}", "ixp_id"),
-                                ("state '{}'", "state")]:
-                    try:
-                        res.append(info.format(connection[k]))
-                    except:
-                        pass
-
-                try:
-                    for iface in connection["if_list"]:
-                        res.append(
-                            "iface on switch_id {}".format(iface["switch_id"])
-                        )
-                except:
-                    pass
-
-            if res:
-                return ", ".join(res)
-            else:
-                return "unknown"
-
-        def process_member(member, clients):
-            if get_item("member_type", member, str, True) == "routeserver":
-                # Member is a route server itself.
-                return
-
-            connection_list = get_item("connection_list", member, list)
-
-            for connection in connection_list:
-                try:
-                    process_connection(member, connection, clients)
-                except EuroIXError as e:
-                    if str(e):
-                        logging.error(
-                            "Error while processing {}: {}".format(
-                                get_descr(member, connection), str(e)
-                            )
-                        )
-                    raise EuroIXError()
-
-        def process_connection(member, connection, clients):
-            check_type(connection, "connection", dict)
-
-            if get_item("ixp_id", connection, int) != ixp_id:
-                return
-
-            # Member has a connection to the selected IXP infrastructure.
-            asnum = get_item("asnum", member, int)
-            name = get_item("name", member, str, True)
-
-            vlan_list = get_item("vlan_list", connection, list, True)
-
-            if vlan_id and not vlan_list:
-                # A specific VLAN has been requested but member does not
-                # have information about VLANs at all.
-                return
-
-            for vlan in vlan_list or []:
-                check_type(vlan, "vlan", dict)
-
-                if vlan_id and \
-                    get_item("vlan_id", vlan, int, True) != vlan_id:
-                    # This VLAN is not the requested one.
-                    continue
-
-                for ip_ver in (4, 6):
-                    ipv4_6 = "ipv{}".format(ip_ver)
-                    ip_info = get_item(ipv4_6, vlan, dict, True)
-
-                    if not ip_info:
-                        continue
-
-                    address = get_item("address", ip_info, str, True)
-
-                    if not address:
-                        continue
-
-                    if routeserver_only:
-                        # Members with routeserver attribute == False
-                        # are excluded.
-                        if get_item("routeserver", ip_info, bool, True) is False:
-                            continue
-
-                    client = new_client(asnum, name)
-                    client["ip"] = address
-
-                    as_macro = get_item("as_macro", ip_info, str, True)
-                    max_prefix = get_item("max_prefix", ip_info, int, True)
-
-                    if as_macro or max_prefix:
-                        client["cfg"] = {
-                            "filtering": {}
-                        }
-                    if as_macro:
-                        client["cfg"]["filtering"]["irrdb"] = {
-                            "as_sets": [as_macro]
-                        }
-                    if max_prefix:
-                        client["cfg"]["filtering"]["max_prefix"] = {
-                            "limit_ipv{}".format(ip_ver): max_prefix
-                        }
-
-                    clients.append(client)
-
-        version = get_item("version", data, str)
-        tested_versions = ClientsFromEuroIXCommand.TESTED_EUROIX_SCHEMA_VERSIONS
-        if version not in tested_versions:
-            logging.warning("The version of the JSON schema of this file ({}) "
-                            "is not one of those tested ({}). Unexpected "
-                            "errors may occurr.".format(
-                                version,
-                                ", ".join(tested_versions)
-                            ))
-
-        ixp_list = get_item("ixp_list", data, list)
-        member_list = get_item("member_list", data, list)
-
-        ixp_found = False
-        for ixp in ixp_list:
-            check_type(ixp, "ixp", dict)
-
-            if get_item("ixp_id", ixp, int) == ixp_id:
-                ixp_found = True
-                break
-
-        if not ixp_found:
-            raise EuroIXError(
-                "IXP ID {} not found".format(ixp_id))
-
-        raw_clients = []
-        for member in member_list:
-            try:
-                check_type(member, "member", dict)
-                process_member(member, raw_clients)
-            except EuroIXError as e:
-                if str(e):
-                    logging.error(
-                        "Error while processing member {}: {}".format(
-                            get_descr(member), str(e)
-                        )                            
-                    )
-                raise EuroIXError()
-
-        return raw_clients
-
-        # TODO: Merge clients with same IRRDB info
-
     def run(self):
-        if self.args.url:
-            try:
-                response = urlopen(self.args.url)
-                raw = response.read().decode("utf-8")
-            except Exception as e:
-                raise EuroIXError(
-                    "Error while retrieving Euro-IX JSON file from {}: {}".format(
-                        self.args.url, str(e)
-                    )
-                )
+        euro_ix = EuroIXMemberList(self.args.url or self.args.input_file)
+
+        if self.args.ixp_id:
+            clients = euro_ix.get_clients(
+                self.args.ixp_id, vlan_id=self.args.vlan_id,
+                routeserver_only=self.args.routeserver_only)
+            res = {"clients": clients}
+
+            yaml.safe_dump(res, self.args.output_file, default_flow_style=False)
         else:
-            raw = self.args.input_file.read()
-
-        try:
-            data = json.loads(raw)
-        except Exception as e:
-            raise EuroIXSchemaError(
-                "Error while processing JSON data: {}".format(str(e))
-            )
-
-        clients = self.clients_from_euroix(
-            data, self.args.ixp_id, vlan_id=self.args.vlan_id,
-            routeserver_only=self.args.routeserver_only)
-        res = {"clients": clients}
-
-        yaml.safe_dump(res, self.args.output_file, default_flow_style=False)
+            euro_ix.print_infrastructure_list(self.args.output_file)
 
         return True